rename link_grammar to link_parser
